Consolidate geocoding functions
