+++ recipes/dep_cats_permits/build.py
@@ -4,59 +4,8 @@ sys.path.insert(0, "..")
 import pandas as pd
 import numpy as np
 import re
-from _helper.geo import get_hnum, get_sname, air_geocode as geocode
+from _helper.geo import get_hnum, get_sname, clean_street, clean_house, clean_boro_name, geocode
 from multiprocessing import Pool, cpu_count
-
-
-def clean_boro(b):
-    if b == "STATENISLAND":
-        b = "STATEN ISLAND"
-    if b not in ["BRONX", "MANHATTAN", "BROOKLYN", "QUEENS", "STATEN ISLAND"]:
-        b = None
-    if b != None:
-        b = b.title()
-    return b
-
-
-def clean_house(s):
-    s = " " if s == None else s
-    s = "" if s[0] == "0" else s
-    s = (
-        re.sub(r"\([^)]*\)", "", s)
-        .replace(" - ", "-")
-        .split("(", maxsplit=1)[0]
-        .split("/", maxsplit=1)[0]
-    )
-    return s
-
-
-def clean_street(s):
-    s = "" if s == None else s
-    s = "JFK INTERNATIONAL AIRPORT" if "JFK" in s else s
-    s = (
-        re.sub(r"\([^)]*\)", "", s)
-        .replace("'", "")
-        .replace("VARIOUS", "")
-        .replace("LOCATIONS", "")
-        .split("(", maxsplit=1)[0]
-        .split("/", maxsplit=1)[0]
-    )
-    return s
-
-
-def parse_streetname(x, n):
-    x = "" if x is None else x
-    if (
-        ("&" in x)
-        | (" AND " in x.upper())
-        | ("CROSS" in x.upper())
-        | ("CRS" in x.upper())
-    ):
-        x = re.split("&| AND | and |CROSS|CRS", x)[n]
-    else:
-        x = ""
-    return x
-
 
 def _import() -> pd.DataFrame:
     url = "https://data.cityofnewyork.us/api/views/f4rp-2kvy/rows.csv"
@@ -101,24 +50,16 @@ def _import() -> pd.DataFrame:
         inplace=True,
     )
 
-    df["borough"] = df.borough.apply(clean_boro)
+    df["borough"] = df.borough.apply(clean_boro_name)
     df["borough"] = np.where(
         (df.streetname.str.contains("JFK")) & (df.borough == None), "Queens", df.borough
     )
-    df["hnum"] = df.housenum.astype(str).apply(clean_house)
-    df["sname"] = df.streetname.astype(str).apply(clean_street)
 
-    df["address"] = df.hnum + " " + df.sname
-
+    df["address"] = df.housenum.astype(str).apply(clean_house) + \
+                    " " + df.streetname.astype(str).apply(clean_street)
     df["hnum"] = df.address.apply(get_hnum)
     df["sname"] = df.address.apply(get_sname)
-
-    df["streetname_1"] = (
-        df["address"].apply(lambda x: parse_streetname(x, 0)).apply(get_sname)
-    )
-    df["streetname_2"] = (
-        df["address"].apply(lambda x: parse_streetname(x, -1)).apply(get_sname)
-    )
+    
     df["status"] = df["status"].apply(lambda x: x.strip())
 
     return df
